gacha_db.go の作成
